Enrich files
